requirements.txt
# requirements.txt
pandas==2.1.4
pyarrow==14.0.2
requests==2.31.0
PyYAML==6.0.1
python-dotenv==1.0.0
